rename filebrowser and refactor out innerfilebrowser
